backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
tensorflow==2.10.1
Pillow==8.4.0
numpy==1.21.2
